compiles and links on both cpu and gpu
